feat(dal): migrate from GORM to sqlc for type-safe query generation
- add sqlc.yaml config with SQLite engine and emit_pointers_for_null_types
- add SQL query files for UpsertIssue and GetLastUpdateTime
- rewrite IssueRepository to use sqlc-generated Queries with explicit transactions
- replace database.NewGormDB with database.NewDB returning *sql.DB via otelsql
- update Migrator to accept *sql.DB directly instead of *gorm.DB
- replace dal.IsNotFound to check sql.ErrNoRows instead of gorm.ErrRecordNotFound
- strip GORM tags from model structs, add TODO(v1-epic) references
- align model field types with SQLite schema (uint64 → int64, int → int64)
- add NOT NULL constraints to issues DDL where columns are never null
- add sqlc generate step to Taskfile
- update Wire bindings and regenerate wire_gen.go
- exclude go-sql-driver/mysql from license-eye dependency check
- remove gorm.io/gorm and glebarez/sqlite dependencies

+++ internal/domain/issue/parser.go
@@ -54,7 +54,7 @@ func (p *Parser) Run(ctx context.Context) error {
 
 	lastUpdate, err := p.lastUpdate(ctx)
 	if err != nil {
-		// SQLLite GORM 1 connection pool. Always permanent
+		// SQLLite 1 connection pool. Always permanent
 		return fmt.Errorf("%w: %w", ErrReadLastUpdate, err)
 	}
 
@@ -110,7 +110,7 @@ func (p *Parser) mapIssueToModel(issue *github.Issue) *model.Issue {
 
 	result := &model.Issue{
 		GithubID:        issue.GetID(),
-		Number:          issue.GetNumber(),
+		Number:          int64(issue.GetNumber()),
 		Title:           issue.GetTitle(),
 		Body:            issue.GetBody(),
 		State:           state,
